fix(dry): Extract shared config loader to eliminate DRY violation
Move the duplicated config file discovery and section loading logic
from infrastructure.py into a shared load_linter_config_section()
utility in shared.py. This eliminates the 41-line duplication flagged
across CLI linter modules.
